// src/app/models/ListaEspera.ts
export interface ListaEspera {
    uuid: string;
    name: string;
    service: string;
    car: string;
    registerHour: Date;
    Timer: string,
    finalized: boolean;
}

export interface Servicio {
    uuid: string;
    name: string;
    service: string;
    car: string;
    finalized: boolean;

}